+++ rust/runnable.rs
@@ -12,7 +12,7 @@ use crate::model::Program;
 use crate::riscv64::RiscV;
 use crate::symbol::Loc;
 use crate::{utils::*, OPT_LEVEL_MASK, OPT_LEVEL_SHIFT};
-use crate::{FASTMATH, SANITIZE, USE_SIMD, USE_THREADS};
+use crate::{FASTMATH, USE_SIMD, USE_THREADS};
 
 use rayon::prelude::*;
 
@@ -99,7 +99,10 @@ impl Application {
 
         let params = vec![0.0; count_params + 1];
 
-        let fastmath = opt & FASTMATH != 0;
+        let fastmath = (opt & FASTMATH != 0)
+            && (!Platform::is_amd64() || Platform::has_avx())
+            && !matches!(ty, CompilerType::AmdSSE);
+
         let mir = prog
             .builder
             .compile_mir(fastmath, Self::opt_level(ty, opt), df)?;
@@ -167,24 +170,8 @@ impl Application {
         })
     }
 
-    fn opt_level(ty: CompilerType, opt: u32) -> u8 {
-        let level: u8 = ((opt & OPT_LEVEL_MASK) >> OPT_LEVEL_SHIFT) as u8;
-
-        if opt & SANITIZE == 0 {
-            return level;
-        }
-
-        match ty {
-            CompilerType::AmdSSE => 0,
-            CompilerType::Native => {
-                if Platform::is_amd64() && !Platform::has_avx() {
-                    0
-                } else {
-                    level
-                }
-            }
-            _ => level,
-        }
+    fn opt_level(_: CompilerType, opt: u32) -> u8 {
+        ((opt & OPT_LEVEL_MASK) >> OPT_LEVEL_SHIFT) as u8
     }
 
     /********************* compile_* functions *************************/